feat: 미완성 entity들 작성
