+++ src/main/java/com/weady/weady/domain/board/entity/BoardGood.java
@@ -1,0 +1,2 @@
+package com.weady.weady.domain.board.entity;public class BoardGood {
+}
