+++ src/main/java/com/weady/weady/domain/tags/entity/ClothesStyleCategory.java
@@ -1,0 +1,2 @@
+package com.weady.weady.domain.tags.entity;public class ClothesStyleCategory {
+}
