+++ src/main/java/com/weady/weady/domain/weather/entity/DailySummary.java
@@ -1,0 +1,2 @@
+package com.weady.weady.domain.weather.entity;public class DailySummary {
+}
